+++ src/Scanner/MethodScanner.php
@@ -451,12 +451,13 @@ class MethodScanner implements ScannerInterface
             if (is_string($token)) {
                 $tokenType    = null;
                 $tokenContent = $token;
-                $tokenLine    = $tokenLine + substr_count(
-                    $lastTokenArray[1],
+                $tokenLine   += substr_count(
+                    $lastTokenArray[1] ?? '',
                     "\n"
                 ); // adjust token line by last known newline count
             } else {
-                list($tokenType, $tokenContent, $tokenLine) = $token;
+                $lastTokenArray = $token;
+                [$tokenType, $tokenContent, $tokenLine] = $token;
             }
 
             return $tokenIndex;